requirements.txt
# Module 9.2: Multi-Hop & Recursive Retrieval
# Core dependencies for multi-hop retrieval with knowledge graphs

# Vector Database
pinecone-client==3.0.3

# Knowledge Graph Database
neo4j==5.15.0

# LLM Providers
openai==1.12.0
anthropic==0.18.1

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.3

# Data Processing
numpy==1.26.4
pandas==2.2.0

# Environment & Config
python-dotenv==1.0.1

# Monitoring (optional)
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
